+++ example/lib/tabs/home_page.dart
@@ -1,5 +1,5 @@
 import 'package:bart/bart/bart_bottombar_actions.dart';
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 
 import 'package:bart/bart.dart';