+++ test/bart_test.dart
@@ -4,7 +4,7 @@ import 'package:bart/bart/bart_scaffold.dart';
 import 'package:bart/bart/widgets/animated_appbar.dart';
 import 'package:bart/bart/widgets/bottom_bar/bottom_bar.dart';
 import 'package:bart/bart/widgets/bottom_bar/styles/bottom_bar_material.dart';
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'components/custom_bottom_bar.dart';