+++ test/components/page_counter.dart
@@ -1,5 +1,5 @@
 import 'package:bart/bart/bart_appbar.dart';
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 
 class PageFakeCounter extends StatefulWidget {
   final ValueNotifier counter = ValueNotifier(1);